Addon Manager: Refactor to extract connection check GUI
